Safari-609.1.11.1/LayoutTests/platform/mac/accessibility/notification-listeners-expected.txt

This tests that a notification listener on an element only listens to that one element, and that a global notification listener listens to all notifications.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Slider
SELECT AXInvalidStatusChanged
GLOBAL AXInvalidStatusChanged on element with role AXRole: AXPopUpButton
SLIDER AXValueChanged
GLOBAL AXValueChanged on element with role AXRole: AXSlider
PASS selectNotificationCount is 1
PASS sliderNotificationCount is 1
PASS globalNotificationCount is 2
PASS successfullyParsed is true

TEST COMPLETE
